Add tag and tag_instance tables
Added tag and tag_instance tables and generated models. Created enum
for types in the tag table and created the same enum in the postgres
SQL database to maintain data integrity. Changed config (in application.
rb) to use sql instead of schema as source of database truth to
accomodate new postgres and rails enum creation.